+++ src/pages/Profile/Profile.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import StandardTab from '@components/Tab/StandardTab';
 import ProfileSection from './Section/ProfileSection';
-import AttendanceTab from './Tab/AttendanceTab';
+import AttendanceTab from './Tab/AttendanceTab/AttendanceTab';
 import BoardTab from './Tab/BoardTab';
 import BookTab from './Tab/BookTab';
 import PointTab from './Tab/PointTab';
